+++ pdoc/doc.py
@@ -1119,7 +1119,7 @@ class Variable(Doc[None]):
                 f"The default value of {self.fullname} matches the {env_var} environment variable. "
                 f"To prevent accidental leakage of secrets, the default value is not displayed. "
                 f"Disable this behavior by setting PDOC_DISPLAY_ENV_VARS=1 as an environment variable.",
-                RuntimeWarning
+                RuntimeWarning,
             )
             return env_var
 
@@ -1146,10 +1146,7 @@ def _environ_lookup():
     """
     A reverse lookup of os.environ. This is a cached function so that it is evaluated lazily.
     """
-    return {
-        value: key
-        for key, value in os.environ.items()
-    }
+    return {value: key for key, value in os.environ.items()}
 
 
 class _PrettySignature(inspect.Signature):