+++ CHANGELOG.md
@@ -8,6 +8,7 @@
    pdoc will now emit a warning and display the variable's name as a placeholder instead.
    This heuristic is meant to prevent accidental leakage of environment secrets and can be disabled by setting
    `PDOC_DISPLAY_ENV_VARS=1`.
+   ([#622](https://github.com/mitmproxy/pdoc/pull/622), @mhils)
 
 ## 2023-09-10: pdoc 14.1.0