+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 <!-- ✨ You do not need to add a pull request reference or an author, this will be added automatically by CI. ✨ -->
 
+ - Add support for `code-block` ReST directives
  - If a variable's value meets certain entropy criteria and matches an environment variable value,
    pdoc will now emit a warning and display the variable's name as a placeholder instead.
    This heuristic is meant to prevent accidental leakage of environment secrets and can be disabled by setting